move tests to top level
