+++ test/socketManager-tests.js
@@ -1,0 +1,2 @@
+const { assert } = require('chai');
+const socketManager = require('../server/socketManager');
